consolidate automation create/edit func
