+++ frontend/src/routes/Governance/policies/EditPolicyAutomation.tsx
@@ -1,8 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { EditMode, useData, useItem } from '@patternfly-labs/react-form-wizard'
 import { PolicyAutomationWizard } from '@patternfly-labs/react-form-wizard/lib/wizards/PolicyAutomation/PolicyAutomationWizard'
-import { AcmToastContext } from '@stolostron/ui-components'
-import { useContext, useMemo } from 'react'
+import { useMemo } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
 import { useRecoilState } from 'recoil'
 import {
@@ -17,13 +16,12 @@ import { SyncEditor } from '../../../components/SyncEditor/SyncEditor'
 import { useTranslation } from '../../../lib/acm-i18next'
 import { NavigationPath } from '../../../NavigationPath'
 import {
-    IResource,
     listAnsibleTowerJobs,
     PolicyAutomation,
-    reconcileResources,
     Secret,
     SubscriptionOperator,
 } from '../../../resources'
+import { handlePolicyAutomationSubmit } from '../common/util'
 import schema from './schemaAutomation.json'
 
 export function WizardSyncEditor() {
@@ -51,7 +49,6 @@ export function EditPolicyAutomation() {
     const { t } = useTranslation()
     const params = useParams<{ namespace: string; name: string }>()
     const { name, namespace } = params
-    const toast = useContext(AcmToastContext)
     const history = useHistory()
     const policies = usePolicies()
     const [secrets] = useRecoilState(secretsState)
@@ -107,50 +104,7 @@ export function EditPolicyAutomation() {
             onCancel={() => history.push(NavigationPath.policies)}
             isAnsibleOperatorInstalled={isOperatorInstalled}
             configMaps={configMaps}
-            onSubmit={(data) => {
-                const resource = data as PolicyAutomation
-                const resources: IResource[] = [resource]
-                if (resource) {
-                    // Copy the cedential to the namespace of the policy
-                    const credToCopy: Secret[] = secrets.filter(
-                        (secret: Secret) =>
-                            secret.metadata.labels?.['cluster.open-cluster-management.io/type'] === 'ans' &&
-                            secret.metadata.name === resource.spec.automationDef.secret
-                    )
-                    const credExists = credToCopy.find(
-                        (cred) => cred.metadata.namespace === resource.metadata.namespace
-                    )
-                    if (!credExists) {
-                        // unshift so secret is created before the PolicyAutomation
-                        resources.unshift({
-                            ...credToCopy[0],
-                            metadata: {
-                                annotations: credToCopy[0].metadata.annotations,
-                                name: credToCopy[0].metadata.name,
-                                namespace: resource.metadata.namespace!,
-                                labels: {
-                                    'cluster.open-cluster-management.io/type': 'ans',
-                                    'cluster.open-cluster-management.io/copiedFromNamespace':
-                                        credToCopy[0].metadata.namespace!,
-                                    'cluster.open-cluster-management.io/copiedFromSecretName':
-                                        credToCopy[0].metadata.name!,
-                                },
-                            },
-                        })
-                    }
-                }
-                return reconcileResources(resources, [currentPolicyAutomation]).then(() => {
-                    if (resource) {
-                        toast.addAlert({
-                            title: t('Policy automation created'),
-                            message: t('{{name}} was successfully created.', { name: resource.metadata?.name }),
-                            type: 'success',
-                            autoClose: true,
-                        })
-                    }
-                    history.push(window.history?.state?.state?.from ?? NavigationPath.policies)
-                })
-            }}
+            onSubmit={(data) => handlePolicyAutomationSubmit (data, secrets, history, t, currentPolicyAutomation)}
             getAnsibleJobsCallback={async (credential: any) => {
                 const host = Buffer.from(credential.data.host || '', 'base64').toString('ascii')
                 const token = Buffer.from(credential.data.token || '', 'base64').toString('ascii')
